timeseries_environment_dockerfull/consumer/requirements.txt
kafka-python==2.0.2
influxdb-client==1.45.0
python-dotenv>=1.0.0